🚚 rename github contribution projector
