--- marketplace-domain/src/contribution/projectors/contribution/mod.rs
@@ -1,6 +1,0 @@
-#[allow(clippy::module_inception)]
-mod github;
-pub use github::GithubProjector as ContributionProjector;
-
-#[cfg(test)]
-mod test;
